requirements/requirements_all.txt
aioredis==1.3.1
amqp==5.1.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
async-timeout==4.0.2
attrdict3==2.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
billiard==3.6.4.0
celery==5.2.7
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cron-descriptor==1.2.35
cryptography==39.0.2
daphne==3.0.2
Django==3.2.18
django-celery-beat==2.5.0
django-cors-headers==3.14.0
django-environ==0.10.0
django-phonenumber-field==7.0.2
django-simple-history==3.3.0
django-timezone-field==5.0
django-webpack-loader==1.8.1
djangorestframework==3.14.0
drf-rw-serializers==1.0.5
drf-spectacular==0.26.1
hiredis==2.2.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
Jinja2==3.1.2
jsonschema==4.17.3
kombu==5.2.4
MarkupSafe==2.1.2
msgpack==1.0.5
mypy-lang==0.5.0
OSlash==0.5.1
phonenumbers==8.13.7
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==23.0.0
pyrsistent==0.19.3
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2022.7
uritemplate==4.1.1
vine==5.0.0
wcwidth==0.2.6
zope.interface==6.0
